make entropy is more Go-y
