+++ go/teams/rename.go
@@ -106,17 +106,20 @@ func RenameSubteam(ctx context.Context, g *libkb.GlobalContext, prevName keybase
 }
 
 func generateRenameSubteamSigForParentChain(g *libkb.GlobalContext, me *libkb.User, signingKey libkb.GenericKey, parentTeam *TeamSigChainState, subteamID keybase1.TeamID, newSubteamName keybase1.TeamName, admin *SCTeamAdmin) (item *libkb.SigMultiItem, err error) {
-	teamSection, err := (SCTeamSection{
+
+	entropy, err := makeSCTeamEntropy()
+	if err != nil {
+		return nil, err
+	}
+
+	teamSection := SCTeamSection{
 		Admin: admin,
 		ID:    (SCTeamID)(parentTeam.GetID()),
 		Subteam: &SCSubteam{
 			ID:   (SCTeamID)(subteamID),
 			Name: (SCTeamName)(newSubteamName.String()),
 		},
-	}).addEntropy()
-
-	if err != nil {
-		return nil, err
+		Entropy: entropy,
 	}
 
 	sigBody, err := RenameSubteamSig(me, signingKey, parentTeam, teamSection)
