+++ go/teams/teams.go
@@ -564,16 +564,18 @@ func (t *Team) deleteSubteam(ctx context.Context) error {
 
 	subteamName := SCTeamName(t.Data.Name.String())
 
-	parentSection, err := (SCTeamSection{
+	entropy, err := makeSCTeamEntropy()
+	if err != nil {
+		return err
+	}
+	parentSection := SCTeamSection{
 		ID: SCTeamID(parentTeam.ID),
 		Subteam: &SCSubteam{
 			ID:   SCTeamID(t.ID),
 			Name: subteamName, // weird this is required
 		},
-		Admin: admin,
-	}).addEntropy()
-	if err != nil {
-		return err
+		Admin:   admin,
+		Entropy: entropy,
 	}
 
 	mr, err := t.G().MerkleClient.FetchRootFromServer(ctx, libkb.TeamMerkleFreshnessForAdmin)
@@ -716,13 +718,16 @@ func (t *Team) postInvite(ctx context.Context, invite SCTeamInvite, role keybase
 		return errors.New("You cannot invite an owner to a team.")
 	}
 
-	teamSection, err := (SCTeamSection{
+	entropy, err := makeSCTeamEntropy()
+	if err != nil {
+		return err
+	}
+
+	teamSection := SCTeamSection{
 		ID:      SCTeamID(t.ID),
 		Admin:   admin,
 		Invites: &invites,
-	}).addEntropy()
-	if err != nil {
-		return err
+		Entropy: entropy,
 	}
 
 	mr, err := t.G().MerkleClient.FetchRootFromServer(ctx, libkb.TeamMerkleFreshnessForAdmin)
